todos los servicios con token ok

+++ Backend/MethodDelphi/MethodFuzzyDelphiWS/src/main/java/com/epn/fd/WS/PersonFacadeREST.java
@@ -7,18 +7,22 @@ package com.epn.fd.WS;
 
 import com.epn.dtos.EmailContainer;
 import com.epn.entities.Person;
+import com.epn.exception.AppException;
 import com.epn.fd.dao.Mail;
 import com.epn.fd.dao.PersonDAO;
+import com.epn.fd.dao.UserDAO;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import java.util.List;
 import javax.ejb.Stateless;
 import javax.inject.Inject;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
+import javax.servlet.http.HttpServletResponse;
 import javax.transaction.Transactional;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
+import javax.ws.rs.HeaderParam;
 import javax.ws.rs.POST;
 import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
@@ -26,6 +30,7 @@ import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
 
 /**
  *
@@ -35,8 +40,11 @@ import javax.ws.rs.core.MediaType;
 @Path("com.epn.entities.person")
 public class PersonFacadeREST extends AbstractFacade<Person> {
 
-    @Inject
+    @Inject()
     PersonDAO personDAO;
+
+    @Inject()
+    UserDAO userDAO;
 
     @PersistenceContext(unitName = "com.epn.fuzzydelphi_MethodFuzzyDelphiWS_war_1.0PU")
     private EntityManager em;
@@ -47,32 +55,46 @@ public class PersonFacadeREST extends AbstractFacade<Person> {
 
     @GET
     @Path("person")
+    @Consumes({MediaType.APPLICATION_JSON})
     @Produces({MediaType.APPLICATION_JSON})
     public String getPerson(
             @QueryParam("from") Integer from,
-            @QueryParam("to") Integer to
+            @QueryParam("to") Integer to,
+            @HeaderParam("authorization") String authString
     ) throws JsonProcessingException {
-
-        return personDAO.getallperson(from, to);
+        String allperson = null;
+        if (userDAO.existToken(authString) == true) {
+            allperson = personDAO.getallperson(from, to);
+        }
+        return allperson;
     }
 
     @GET
     @Path("personbycode")
     @Produces({MediaType.APPLICATION_JSON})
     public String getPersonbycode(
-            @QueryParam("codeperson") Long codeperson
+            @QueryParam("codeperson") Long codeperson,
+            @HeaderParam("authorization") String authString
     ) throws JsonProcessingException {
-
-        return personDAO.getpersonbycode(codeperson);
+        String personbycode = null;
+        if (userDAO.existToken(authString) == true) {
+            personbycode = personDAO.getpersonbycode(codeperson);
+        }
+        return personbycode;
     }
 
     @POST
     @Path("save")
     @Transactional
     @Consumes({MediaType.APPLICATION_JSON})
-    public Person saveperson(Person person) {
+    public Person saveperson(
+            Person person,
+            @HeaderParam("authorization") String authString) {
+        Person personsaved = new Person();
+        if (userDAO.existToken(authString) == true) {
+            personsaved = personDAO.savePerson(person);
+        }
 
-        Person personsaved = personDAO.savePerson(person);
         return personsaved;
     }
 
@@ -86,9 +108,10 @@ public class PersonFacadeREST extends AbstractFacade<Person> {
     @POST
     @Path("delete")
     @Consumes({MediaType.APPLICATION_JSON})
-    public void delete(List<Person> idlist) {
-
-        personDAO.deletePerson(idlist);
+    public void delete(List<Person> idlist,@HeaderParam("authorization") String authString) {
+        if (userDAO.existToken(authString) == true) {
+           personDAO.deletePerson(idlist);
+        }
     }
 
     @PUT
